@@ -1,61 +1,91 @@
 {
   "name": "datashare-web",
   "version": "0.1.0",
+  "type": "module",
   "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
+    "set-env": "../scripts/sync-env-web.sh",
+    "start": "npm run set-env && ng serve --port $WEB_PORT",
+    "build": "npm run set-env && ng build ",
+    "watch": "npm run set-env && ng build --watch --configuration development",
+    "test": "vitest",
+    "test:ui": "vitest --ui",
+    "test:coverage": "vitest --coverage",
+    "test:unit": "vitest --config=config/vitest.unit.config.ts",
+    "test:unit:ui": "vitest --config=config/vitest.unit.config.ts --ui",
+    "test:unit:coverage": "vitest --config=config/vitest.unit.config.ts --coverage",
+    "test:unit:run": "vitest run --config=config/vitest.unit.config.ts",
+    "test:integration": "vitest --config=config/vitest.integration.config.ts",
+    "test:integration:ui": "vitest --config=config/vitest.integration.config.ts --ui",
+    "test:integration:coverage": "vitest --config=config/vitest.integration.config.ts --coverage",
+    "test:integration:run": "vitest run --config=config/vitest.integration.config.ts",
+    "test:e2e": "cypress open",
+    "test:e2e:run": "cypress run",
+    "test:e2e:headless": "cypress run --headless",
+    "test:e2e:chrome": "cypress run --browser chrome",
+    "test:e2e:firefox": "cypress run --browser firefox",
+    "test:all": "npm run test:unit:run && npm run test:integration:run && npm run test:e2e:run",
+    "test:ci": "npm run test:unit:run && npm run test:integration:run && npm run test:e2e:headless",
     "lint": "ng lint",
-    "prettier": "prettier \"src/**/*.ts\" --write",
-    "prettier:check": "prettier \"src/**/*.ts\" --check",
-    "watch": "ng build --watch --configuration development",
-    "test-unit": "jest --config tests/config/jest.config.ts --coverage --passWithNoTests",
-    "test-it": "jest --config tests/config/jest.integration.config.ts --runInBand --passWithNoTests",
-    "test-e2e": "cypress run --config-file tests/config/cypress.config.ts || npx playwright test"
-  },
-  "prettier": {
-    "printWidth": 100,
-    "singleQuote": true,
-    "overrides": [
-      {
-        "files": "*.html",
-        "options": {
-          "parser": "angular"
-        }
-      }
-    ]
+    "format": "prettier --write \"src/**/*.{ts,html,scss,css,json}\"",
+    "format:check": "prettier --check \"src/**/*.{ts,html,scss,css,json}\"",
+    "typedoc": "typedoc src/index.ts src/app/ --out ../../docs/typedoc",
+    "docs": "npm run typedoc"
   },
   "private": true,
   "packageManager": "npm@11.7.0",
   "dependencies": {
-    "@angular/common": "^21.0.0",
-    "@angular/compiler": "^21.0.0",
-    "@angular/core": "^21.0.0",
-    "@angular/forms": "^21.0.0",
-    "@angular/platform-browser": "^21.0.0",
-    "@angular/router": "^21.0.0",
-    "dotenv": "^17.2.3",
-    "rxjs": "~7.8.0",
-    "tslib": "^2.3.0"
+    "@angular/animations": "^21.1.2",
+    "@angular/cdk": "^21.1.2",
+    "@angular/common": "^21.1.2",
+    "@angular/compiler": "^21.1.2",
+    "@angular/core": "^21.1.2",
+    "@angular/forms": "^21.1.2",
+    "@angular/material": "^21.1.2",
+    "@angular/platform-browser": "^21.1.2",
+    "@angular/platform-browser-dynamic": "^21.1.2",
+    "@angular/router": "^21.1.2",
+    "rxjs": "^7.8.2",
+    "zone.js": "^0.16.0"
   },
   "devDependencies": {
-    "@angular/build": "^21.0.5",
-    "@angular/cli": "^21.0.5",
-    "@angular/compiler-cli": "^21.0.0",
-    "@playwright/test": "^1.57.0",
-    "@types/dotenv": "^6.1.1",
-    "@types/node": "^25.0.7",
-    "angular-eslint": "21.1.0",
-    "cypress": "^15.8.2",
-    "eslint": "^9.39.1",
+    "@analogjs/vite-plugin-angular": "^2.2.3",
+    "@analogjs/vitest-angular": "^2.2.3",
+    "@angular-eslint/builder": "^21.2.0",
+    "@angular-eslint/eslint-plugin": "^21.2.0",
+    "@angular-eslint/eslint-plugin-template": "^21.2.0",
+    "@angular-eslint/schematics": "^21.2.0",
+    "@angular-eslint/template-parser": "^21.2.0",
+    "@angular/build": "^21.1.2",
+    "@angular/cli": "^21.1.2",
+    "@angular/compiler-cli": "^21.1.2",
+    "@badeball/cypress-cucumber-preprocessor": "^24.0.0",
+    "@bahmutov/cypress-esbuild-preprocessor": "^2.2.8",
+    "@eslint/js": "^9.39.2",
+    "@testing-library/angular": "^19.0.0",
+    "@types/jsdom": "^27.0.0",
+    "@types/node": "^25.1.0",
+    "@typescript-eslint/eslint-plugin": "^8.54.0",
+    "@typescript-eslint/parser": "^8.54.0",
+    "@vitest/coverage-v8": "^4.0.18",
+    "@vitest/ui": "^4.0.18",
+    "angular-eslint": "^21.2.0",
+    "cypress": "^15.9.0",
+    "dotenv": "^17.2.3",
+    "eslint": "^9.39.2",
     "eslint-config-prettier": "^10.1.8",
-    "jest": "^30.2.0",
-    "jsdom": "^27.1.0",
-    "playwright": "^1.57.0",
-    "prettier": "^3.8.0",
-    "ts-jest": "^29.4.6",
-    "typescript": "~5.9.2",
-    "typescript-eslint": "8.47.0",
-    "vitest": "^4.0.8"
+    "eslint-plugin-prettier": "^5.5.5",
+    "jsdom": "^27.4.0",
+    "prettier": "^3.8.1",
+    "typedoc": "^0.28.16",
+    "typescript": "^5.9.3",
+    "typescript-eslint": "^8.54.0",
+    "vite": "^7.3.1",
+    "vitest": "^4.0.18"
+  },
+  "cypress-cucumber-preprocessor": {
+    "stepDefinitions": "src/test/e2e/steps/**/*.ts",
+    "messages": {
+      "enabled": true
+    }
   }
 }
