@@ -1,3 +1,33 @@
 import { Routes } from '@angular/router';
+import { Home } from './pages/home/home';
+import { Register } from './pages/register/register';
+import { Login } from './pages/login/login';
+import { Upload } from './pages/upload/upload';
+import { Files } from './pages/files/files';
 
-export const routes: Routes = [];
+export const routes: Routes = [
+  {
+    path: '',
+    component: Home,
+  },
+  {
+    path: 'home',
+    component: Home,
+  },
+  {
+    path: 'register',
+    component: Register,
+  },
+  {
+    path: 'login',
+    component: Login,
+  },
+  {
+    path: 'upload',
+    component: Upload,
+  },
+  {
+    path: 'files',
+    component: Files,
+  },
+];
